Add files via upload
This project is a classic Snake Game implemented in Java. The Snake Game is a popular retro game where the player controls a snake that continuously moves around the screen. The objective of the game is to eat as much food as possible while avoiding collisions with the walls of the game area and the snake's own body. The snake grows longer each time it consumes food, making the game progressively more challenging.

The game features a simple graphical user interface using Java's Swing library, providing a visually engaging experience for players. It is designed to be easy to play, yet challenging to master, making it suitable for players of all ages.